# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fetch — AI-powered food discovery and nutrition management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fridge:
    get:
      tags: [Fridge]
      operationId: listFridgeItems
      summary: List items in MyFridge
      responses:
        '200':
          description: Items loaded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FridgeItem'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [Fridge]
      operationId: createFridgeItem
      summary: Add an item to MyFridge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FridgeItemInput'
      responses:
        '201':
          description: Item created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FridgeItem'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /fridge/{id}:
    patch:
      tags: [Fridge]
      operationId: updateFridgeItem
      summary: Update a MyFridge item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FridgeItemUpdate'
      responses:
        '200':
          description: Item updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FridgeItem'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Fridge]
      operationId: deleteFridgeItem
      summary: Delete a MyFridge item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted.
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /fridge/from-image:
    post:
      tags: [Fridge]
      operationId: addFridgeItemsFromImage
      summary: Detect ingredients from a photo and add them to MyFridge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanInput'
      responses:
        '200':
          description: Items detected and added.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FridgeItem'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: AI provider error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /fridge/recipes:
    post:
      tags: [Fridge]
      operationId: generateFridgeRecipes
      summary: Generate AI recipes from current MyFridge contents and user memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeRequest'
      responses:
        '200':
          description: Recipes generated.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: AI provider error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ScanInput:
      type: object
      required: [imageObjectPath]
      properties:
        imageObjectPath:
          type: string
          description: Object path returned from `POST /storage/uploads/request-url` after the client PUTs the photo to GCS.
          minLength: 1
        note:
          type: string
          description: Optional user note to bias analysis (e.g. "lunch portion").
...
    FridgeItem:
      type: object
      required:
        - id
        - userId
        - name
        - quantity
        - category
        - addedAt
      properties:
        id: { type: string, format: uuid }
        userId: { type: string }
        name: { type: string }
        quantity: { type: string }
        category:
          type: string
          description: e.g. produce, dairy, meat, pantry, beverage.
        expiresAt: { type: [string, 'null'], format: date-time }
        notes: { type: [string, 'null'] }
        addedAt: { type: string, format: date-time }
    FridgeItemInput:
      type: object
      required: [name, quantity, category]
      properties:
        name: { type: string, minLength: 1 }
        quantity: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        expiresAt: { type: [string, 'null'], format: date-time }
        notes: { type: [string, 'null'] }
    FridgeItemUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        quantity: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        expiresAt: { type: [string, 'null'], format: date-time }
        notes: { type: [string, 'null'] }
    RecipeRequest:
      type: object
      properties:
        cuisine:
          type: string
          description: Optional cuisine preference (e.g. "italian", "thai").
        maxMinutes:
          type: integer
          minimum: 5
          maximum: 240
          description: Maximum total cook time in minutes.
    Recipe:
      type: object
      required:
        - id
        - title
        - description
        - estimatedMinutes
        - servings
        - calories
        - usedIngredients
        - missingIngredients
        - steps
        - tags
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        estimatedMinutes: { type: integer }
        servings: { type: integer }
        calories: { type: integer }
        usedIngredients:
          type: array
          items: { type: string }
        missingIngredients:
          type: array
          items: { type: string }
        steps:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
...