# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fetch — AI-powered food discovery and nutrition management
 * OpenAPI spec version: 0.1.0
 */

export interface ScanInput {
  /**
   * Object path returned from `POST /storage/uploads/request-url` after the client PUTs the photo to GCS.
   * @minLength 1
   */
  imageObjectPath: string;
  /** Optional user note to bias analysis (e.g. "lunch portion"). */
  note?: string;
}